+++ src/app/components/logoutButton/logout.css
@@ -1,0 +1,7 @@
+button{
+    padding:12px 30px;
+    color: #fff;
+    background-color: #0866FF;
+    border: none;
+    border-radius: 10px;
+}
